fix(guard): extract frozen shim helpers for restore CI ratchets
Keep oversized modules at their reviewed sizes, regenerate the command decision-diff fixture, and split frozen Core launch plus Cloud remaining-step parsing into dedicated modules.
