+++ dashboard/src/guard-types.ts
@@ -1044,24 +1044,11 @@ export type PackageFirewallActionResponse = {
   entitlement: PackageFirewallEntitlement;
 };
 
-export type SupplyChainRepairStepFailure = {
-  step: "package_shims" | "runtime_activation" | "intelligence_sync";
-  message: string;
-};
-
-export type SupplyChainRepairRemainingStep = {
-  step: "intelligence_sync";
-  message: string;
-  action: "connect";
-};
-
-export type SupplyChainRepairResult = {
-  repaired: boolean;
-  completed_steps: string[];
-  failed_steps: SupplyChainRepairStepFailure[];
-  remaining_steps: SupplyChainRepairRemainingStep[];
-  message: string;
-};
+export type {
+  SupplyChainRepairRemainingStep,
+  SupplyChainRepairResult,
+  SupplyChainRepairStepFailure,
+} from "./supply-chain-repair-types";
 
 export type SupplyChainAuditDecision = "allow" | "monitor" | "warn" | "ask" | "block";
