+++ src/codex_plugin_scanner/guard/daemon/static/assets/chunks/supply-chain-hub-workspace.js
@@ -1473,12 +1473,39 @@ const IDLE_SUPPLY_CHAIN_FIX_ALL_STATE = {
   phase: "idle",
   message: null,
   completedSteps: [],
-  failedSteps: [],
-  remainingAction: null,
-  remainingSteps: []
+  failedSteps: []
 };
+function supplyChainFixAllWorkingState() {
+  return {
+    phase: "working",
+    message: "Repairing package tools and turning on routing…",
+    completedSteps: [],
+    failedSteps: []
+  };
+}
 function supplyChainFixAllNeedsCloudConnect(state) {
   return state.remainingAction === "connect" && state.failedSteps.length === 0;
+}
+function supplyChainFixAllStateFromRepair(result) {
+  const remainingAction = result.remaining_steps.some((step) => step.action === "connect") ? "connect" : null;
+  return {
+    phase: result.repaired ? "success" : "incomplete",
+    message: result.message,
+    completedSteps: result.completed_steps,
+    failedSteps: result.failed_steps.map((failure) => failure.message),
+    remainingAction,
+    remainingSteps: result.remaining_steps.map((step) => step.message)
+  };
+}
+function supplyChainFixAllConnectState(phase, message, remainingSteps = []) {
+  return {
+    phase,
+    message,
+    completedSteps: [],
+    failedSteps: [],
+    remainingAction: "connect",
+    remainingSteps
+  };
 }
 function supplyChainFixAllButtonLabel(phase, remainingAction = null, failedCount = 0) {
   if (phase === "working") return "Fixing…";
@@ -1833,26 +1860,23 @@ const PackageFirewallPanel = reactExports.forwardRef(function PackageFirewallPan
   }, [onStateChanged, refreshAfterOp]);
   const beginFixAllConnectRecovery = reactExports.useCallback(async () => {
     setResumeFixAllAfterConnect(true);
-    onFixAllStateChange?.({
-      phase: "connecting",
-      message: "Finish Guard Cloud sign-in. Repair will resume here automatically.",
-      completedSteps: [],
-      failedSteps: [],
-      remainingAction: "connect",
-      remainingSteps: []
-    });
+    onFixAllStateChange?.(
+      supplyChainFixAllConnectState(
+        "connecting",
+        "Finish Guard Cloud sign-in. Repair will resume here automatically."
+      )
+    );
     const started = await handleStartConnect();
     if (started) return;
     setResumeFixAllAfterConnect(false);
     repairNeedsCloudConnectRef.current = true;
-    onFixAllStateChange?.({
-      phase: "error",
-      message: "Guard Cloud sign-in could not start. Connect again to continue.",
-      completedSteps: [],
-      failedSteps: [],
-      remainingAction: "connect",
-      remainingSteps: ["Guard Cloud sign-in could not start."]
-    });
+    onFixAllStateChange?.(
+      supplyChainFixAllConnectState(
+        "error",
+        "Guard Cloud sign-in could not start. Connect again to continue.",
+        ["Guard Cloud sign-in could not start."]
+      )
+    );
   }, [handleStartConnect, onFixAllStateChange]);
   const handleFixAll = reactExports.useCallback(
     async (credentials) => {
@@ -1861,29 +1885,15 @@ const PackageFirewallPanel = reactExports.forwardRef(function PackageFirewallPan
         await beginFixAllConnectRecovery();
         return;
       }
-      onFixAllStateChange?.({
-        phase: "working",
-        message: "Repairing package tools and turning on routing…",
-        completedSteps: [],
-        failedSteps: [],
-        remainingAction: null,
-        remainingSteps: []
-      });
+      onFixAllStateChange?.(supplyChainFixAllWorkingState());
       setPendingOp({ op: "fix_all", manager: null });
       try {
         const result = await repairSupplyChainProtection(credentials);
         await refreshAfterOp();
         await onStateChanged?.();
-        const remainingAction = result.remaining_steps.some((step) => step.action === "connect") ? "connect" : null;
-        repairNeedsCloudConnectRef.current = remainingAction === "connect" && result.failed_steps.length === 0;
-        onFixAllStateChange?.({
-          phase: result.repaired ? "success" : "incomplete",
-          message: result.message,
-          completedSteps: result.completed_steps,
-          failedSteps: result.failed_steps.map((failure) => failure.message),
-          remainingAction,
-          remainingSteps: result.remaining_steps.map((step) => step.message)
-        });
+        const nextState = supplyChainFixAllStateFromRepair(result);
+        repairNeedsCloudConnectRef.current = supplyChainFixAllNeedsCloudConnect(nextState);
+        onFixAllStateChange?.(nextState);
       } catch (error) {
         if (credentials === void 0 && isApprovalGateRequiredError(error)) {
           await resolveApprovalGate();
@@ -1931,16 +1941,15 @@ const PackageFirewallPanel = reactExports.forwardRef(function PackageFirewallPan
       const connectFailed = panelLoad.data.connect_flow?.state === "failed";
       setResumeFixAllAfterConnect(false);
       repairNeedsCloudConnectRef.current = true;
-      onFixAllStateChange?.({
-        phase: "error",
-        message: connectFailed ? panelLoad.data.connect_flow?.detail || "Guard Cloud sign-in did not finish. Connect again to continue." : "Guard Cloud sign-in did not grant package protection access. Connect again or review plan access.",
-        completedSteps: [],
-        failedSteps: [],
-        remainingAction: "connect",
-        remainingSteps: [
-          connectFailed ? "Guard Cloud sign-in did not finish." : "Package protection access is still unavailable."
-        ]
-      });
+      onFixAllStateChange?.(
+        supplyChainFixAllConnectState(
+          "error",
+          connectFailed ? panelLoad.data.connect_flow?.detail || "Guard Cloud sign-in did not finish. Connect again to continue." : "Guard Cloud sign-in did not grant package protection access. Connect again or review plan access.",
+          [
+            connectFailed ? "Guard Cloud sign-in did not finish." : "Package protection access is still unavailable."
+          ]
+        )
+      );
       return;
     }
     if (!panelLoad.data.entitlement.allowed) return;
