+++ tests/fixtures/guard-command-corpus/decision-diff-report.json
@@ -54,8 +54,9 @@
       "source-1e1bef156ccecc6aa514b161": "0754f3d94389eb24d2462ee66ba3cbdb991a6f915ae9964c60f4c98748ddf93d",
       "source-1fd0cc4dfab87191da06f18d": "2c2230547b2039d6589b425dafa4620600e24016880675775ee9585e7bf20fc8",
       "source-202529ee38d88bb604dec657": "0e384c4ab18ccb9a4db26b9b29af793a4d5132ae6092aca5603186172d7d5500",
+      "source-2038f0de57a835fb492ef16a": "93140e99df4f899319f853ddef4f121bcb24e76b6556046957e3ac2454ce75c4",
       "source-2131ffc4bad823c069aadaa8": "534232af6dc97a8486104ebeccddf473518352a05497e962a4ca0e278e94ab61",
-      "source-21569bf21f9c57722bf96fc8": "931c1bc39c896490ce116b5ce029bcda857d4cfb798ba0775a788b6219f0fa91",
+      "source-21569bf21f9c57722bf96fc8": "8c790e337017a5b7cafcc4149f418194e50b7592827555a788270f8c134f9796",
       "source-21fb68ac6d21e5b5ddebff04": "c02fba8df5fa06594398cb4694333fc799e530b1a197b9651c2ba45856c543c0",
       "source-22b4a78a9b6077c0a8c16853": "2e2f9fb2d4095af815ac847df521c2a986bf1b6f6468db08ab2f2dd28a7aa1cd",
       "source-22d03d37b90d3ea5b13cd269": "4e3a2bb4270835abad1b30c62c96019e352b962bd42252b6b93058cc5cd867b4",
@@ -200,7 +201,7 @@
       "source-94c910284bb80843ed2dd5f8": "ad51bd039869d4c4624ffdebce9e0be83ba04786d664ec3499abe151f2d43a88",
       "source-95d33f0393508d423a92e8c8": "1a917869da271fabf6227d5a3f6d514d453f15a9077969426825980cdede87a8",
       "source-95d5338c4781f9ff99292efa": "8c347a52222793ab8696167700ae111957653871794ed645207ded70d2a94011",
-      "source-9958bb236bc9db6bc926d4a2": "76adce182912087f7fad7aa1d9663931167628a0e9521c9a5b6b336e8c5299ac",
+      "source-9958bb236bc9db6bc926d4a2": "d80ee25644930282e7b90fc264413e97be79ba3069bb77c481d86a698b0012a6",
       "source-9b042019b4ef949d0e2580bd": "7efb08e9397199cf5751556efcb5efebcd4bc557a1b65b7ae22ee63b849d79d6",
       "source-9be635c8eff62b9e453a4b07": "22cde8189f6c94ef83249c24d2dd7e4f944dc29901055989e7aa1f9f0642c4ef",
       "source-9bf21c89e8dff6638b4092d6": "3900de0b921dce9deac0424228f22ccd1febb021a9db8f3f1d7f0d2c75905846",
@@ -208,7 +209,7 @@
       "source-9df5813f3e985972060c0e37": "57d26320b6fbaf43c7e409654198a2369ee1cb2668e8fa30bcc80f9d6a0fe682",
       "source-9e3a96c5287017c496886547": "a7c82021dcabd52c6f506764d3b4027d24e633d62d2465d03c5b6c109b808405",
       "source-9ede03d120a175084586751a": "b1949b71cd9e81935a3bb51b26ce564b9a496db0f83986114023f277bc510256",
-      "source-9f0568215682047e5f813120": "4d20fbbea469478da9f650cd69ba9bf3b08d51442a769e3b8cd40e42dd0416d7",
+      "source-9f0568215682047e5f813120": "1a8b883c67f3def065080f11276857b634e8e63fe29596bdd8ef0aba0752b397",
       "source-9f4463a8839e64022c7ae1d4": "d13e2f78105c69e5c83db06c207a89a6dea216632b9cb239d2eb81bd7ad35cf3",
       "source-9fffbb9f42c50575e09d4a0e": "56a69f517de93b42382d0af7183e6148f3950562add1b9761f9fa43fd3e295ac",
       "source-a2294ac817357a5e5c7af5f2": "aea88653f5e72c4fb193a80debf7954552086174811a6ac1cdd8f09c8f790569",
