+++ tests/guard_command_decision_diff.py
@@ -98,6 +98,7 @@ _EVIDENCE_SOURCE_PATHS: Final = (
     REPO_ROOT / "src" / "codex_plugin_scanner" / "guard" / "contained_package_script_execution.py",
     REPO_ROOT / "src" / "codex_plugin_scanner" / "guard" / "contained_workspace_write_execution.py",
     REPO_ROOT / "src" / "codex_plugin_scanner" / "guard" / "package_shim_gate.py",
+    REPO_ROOT / "src" / "codex_plugin_scanner" / "guard" / "package_shim_frozen.py",
     REPO_ROOT / "src" / "codex_plugin_scanner" / "guard" / "shims.py",
     *(REPO_ROOT / "src" / "codex_plugin_scanner" / "guard" / "runtime").glob("*.py"),
     *REPO_ROOT.joinpath("tests").glob("guard_command_corpus*.py"),
