+++ tests/test_guard_command_decision_diff.py
@@ -71,6 +71,7 @@ def test_report_is_exactly_reproducible_and_source_bound() -> None:
         "src/codex_plugin_scanner/guard/contained_package_script_execution.py",
         "src/codex_plugin_scanner/guard/contained_workspace_write_execution.py",
         "src/codex_plugin_scanner/guard/package_shim_gate.py",
+        "src/codex_plugin_scanner/guard/package_shim_frozen.py",
         "src/codex_plugin_scanner/guard/shims.py",
         "tests/test_guard_contained_package_script_execution.py",
         "tests/test_guard_contained_workspace_write_contract.py",
